refactor: move settings to featur eoriented structure
