+++ src/lib/features/settings/setting-service.e2e.test.ts
@@ -1,14 +1,16 @@
-import SettingService from '../../../lib/services/setting-service.js';
-import { createTestConfig } from '../../config/test-config.js';
-import dbInit, { type ITestDb } from '../helpers/database-init.js';
-import type { IUnleashStores } from '../../../lib/types/stores.js';
+import SettingService from './setting-service.js';
+import { createTestConfig } from '../../../test/config/test-config.js';
+import dbInit, {
+    type ITestDb,
+} from '../../../test/e2e/helpers/database-init.js';
+import type { IUnleashStores } from '../../types/stores.js';
 import {
     SETTING_CREATED,
     SETTING_DELETED,
     SETTING_UPDATED,
-} from '../../../lib/events/index.js';
-import { createEventsService } from '../../../lib/features/index.js';
-import { TEST_AUDIT_USER } from '../../../lib/types/index.js';
+} from '../../events/index.js';
+import { createEventsService } from '../index.js';
+import { TEST_AUDIT_USER } from '../../types/index.js';
 
 let stores: IUnleashStores;
 let db: ITestDb;
