+++ src/lib/features/settings/settings-store-type.ts
@@ -1,4 +1,4 @@
-import type { Store } from './store.js';
+import type { Store } from '../../types/stores/store.js';
 
 export interface ISettingInsert {
     name: string;
