+++ src/lib/routes/admin-api/user-admin.ts
@@ -9,7 +9,7 @@ import type { IUnleashConfig, RoleName } from '../../types/index.js';
 import type { IUnleashServices } from '../../services/index.js';
 import type ResetTokenService from '../../services/reset-token-service.js';
 import type { IAuthRequest } from '../unleash-types.js';
-import type SettingService from '../../services/setting-service.js';
+import type SettingService from '../../features/settings/setting-service.js';
 import type { IUser } from '../../types/index.js';
 import { anonymise } from '../../util/anonymise.js';
 import type { OpenApiService } from '../../services/openapi-service.js';
