+++ src/lib/services/user-service.test.ts
@@ -11,8 +11,8 @@ import SessionService from './session-service.js';
 import FakeSessionStore from '../../test/fixtures/fake-session-store.js';
 import User from '../types/user.js';
 import FakeResetTokenStore from '../../test/fixtures/fake-reset-token-store.js';
-import SettingService from './setting-service.js';
-import FakeSettingStore from '../../test/fixtures/fake-setting-store.js';
+import SettingService from '../features/settings/setting-service.js';
+import FakeSettingStore from '../features/settings/fake-setting-store.js';
 import { extractAuditInfoFromUser } from '../util/index.js';
 import { createFakeEventsService } from '../features/index.js';
 import { vi, expect, test, describe, beforeEach } from 'vitest';
