+++ src/lib/services/user-service.ts
@@ -31,7 +31,7 @@ import {
 } from '../types/index.js';
 import type { IUserStore } from '../types/index.js';
 import { RoleName } from '../types/model.js';
-import type SettingService from './setting-service.js';
+import type SettingService from '../features/settings/setting-service.js';
 import {
     type SimpleAuthSettings,
     simpleAuthSettingsKey,
